requirements.txt
pymorphy3>=2.0.3
pymorphy2>=0.9.1
pdfminer.six>=20231228
python-docx>=0.8.11
docx2txt>=0.8
